requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
streamlit>=1.12.0
crewai>=0.11.0
